server now working with typescript'
