+++ mobile-browser-based-version/server/package.json
@@ -14,7 +14,6 @@
     "@tensorflow-models/mobilenet": "2.1.0",
     "@tensorflow/tfjs": "^3.13.0",
     "@tensorflow/tfjs-node": "^3.13.0",
-    "class-transformer": "^0.5.1",
     "cors": "^2.8.5",
     "express": "^4.17.2",
     "http-proxy-middleware": "^2.0.1",
@@ -22,7 +21,6 @@
     "msgpack-lite": "^0.1.26",
     "node-fetch": "^3.2.0",
     "peer": "^0.6.1",
-    "reflect-metadata": "^0.1.13",
     "ts-node": "^10.4.0"
   },
   "nodemonConfig": {
