+++ mobile-browser-based-version/server/src/helpers/helpers.ts
@@ -6,9 +6,9 @@
 export function makeID(IDLength) {
   return () => {
     let result = '';
-    let characters =
+    const characters =
       'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-    let charactersLength = characters.length;
+    const charactersLength = characters.length;
     for (let i = 0; i < IDLength; i++) {
       result += characters.charAt(Math.floor(Math.random() * charactersLength));
     }
