+++ mobile-browser-based-version/server/src/router/router.ts
@@ -1,17 +1,13 @@
 import express from 'express';
 import _ from 'lodash';
 import * as handlers from '../logic/federated/handlers';
-import { tasks, writeNewTask, initTasks } from '../tasks/tasks';
+import { writeNewTask, getTasks } from '../tasks/tasks_helper';
 import { ExpressPeerServer } from 'peer';
 import { makeID } from '../helpers/helpers';
 import { createProxyMiddleware } from 'http-proxy-middleware';
 import * as config from '../../server.config';
 
-// import { hello_world } from '../tasks/hello';
-
-// hello_world();
-
-initTasks();
+const tasks = getTasks(config);
 
 // General tasks routes
 const tasksRouter = express.Router();
@@ -20,7 +16,7 @@ tasksRouter.get('/:task/:file', handlers.getLatestModel);
 // POST method route (task-creation-form)
 tasksRouter.post('/', function (req, res) {
   const newTask = req.body,
-    newPort = config.START_TASK_PORT + tasks?.length();
+    newPort = config.START_TASK_PORT + tasks?.length;
   if (newTask['taskID'] in tasks)
     console.log('Cannot add new task (key is already defined in Tasks.json)');
   else {
@@ -73,7 +69,7 @@ const decentralisedRouter = express.Router();
  */
 const ports = _.range(
   config.START_TASK_PORT,
-  config.START_TASK_PORT + tasks?.length()
+  config.START_TASK_PORT + tasks?.length
 );
 const createTaskServer = (task, port) => {
   /**
@@ -105,8 +101,7 @@ const createTaskServer = (task, port) => {
     })
   );
 };
-console.log('Router.ts tasks:', tasks);
-// _.forEach(_.zip(tasks?.tasks, ports), _.spread(createTaskServer));
+_.forEach(_.zip(tasks, ports), _.spread(createTaskServer));
 
 decentralisedRouter.use('/tasks', tasksRouter);
 decentralisedRouter.get('/', (req, res) => res.send('DeAI server'));
