+++ mobile-browser-based-version/server/src/run_server.ts
@@ -1,7 +1,6 @@
 import models from './tasks/models';
 import express from 'express';
 import cors from 'cors';
-import 'reflect-metadata';
 
 import { federatedRouter, decentralisedRouter } from './router/router';
 import * as config from '../server.config';
